Validate against creating a duplicate published edition
It shouldn't be possible to publish an edition of a guide with the same version number. The version number should be incremented for an additional edition/version.

This validation is being added as a result of this happening in production. It was possible to replicate this issue by having a guide in a 'ready' state open in two tabs. Clicking 'Publish' in one tab and then clicking 'Save' (to make a new draft) in the other tab would create a draft without incrementing the version number.

This is pretty rare for the moment so this validation should suffice for the moment.